// frontend/src/utils/dto.types.ts
export type RoomDTO = {
    _id: string;
    title: string;
    address: string;
    price: number;
    images: string[];
    area: number;
    shortDescription: string;
    fullDescription: string;
    views: number;
    contactViews: number;
    status: string;
    owner: string;
    isActive: boolean;
    createdAt: string;
    updatedAt: string;
}